+++ zoe_api/rest_api/userinfo.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2016, Daniele Venzano
+# Copyright (c) 2016, Quang-Nhat Hoang-Xuan
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,16 +13,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""The Discovery API endpoint."""
+"""The Info API endpoint."""
 
 from tornado.web import RequestHandler
-
+from zoe_api.rest_api.utils import get_auth, catch_exceptions, manage_cors_headers
 from zoe_api.api_endpoint import APIEndpoint  # pylint: disable=unused-import
-from zoe_api.rest_api.utils import catch_exceptions, manage_cors_headers
 
 
-class DiscoveryAPI(RequestHandler):
-    """The Discovery API endpoint."""
+class UserInfoAPI(RequestHandler):
+    """The UserInfo API endpoint."""
 
     def initialize(self, **kwargs):
         """Initializes the request handler."""
@@ -32,23 +31,20 @@ class DiscoveryAPI(RequestHandler):
         """Set up the headers for enabling CORS."""
         manage_cors_headers(self)
 
+    @catch_exceptions
     def options(self):
         """Needed for CORS."""
         self.set_status(204)
         self.finish()
 
     @catch_exceptions
-    def get(self, execution_id: int, service_group: str):
+    def get(self):
         """HTTP GET method."""
-        self.api_endpoint.execution_by_id(0, 'admin', execution_id)
-        if service_group != 'all':
-            services = self.api_endpoint.service_list(0, 'admin', service_group=service_group, execution_id=execution_id)
-        else:
-            services = self.api_endpoint.service_list(0, 'admin', execution_id=execution_id)
+        uid, role = get_auth(self)
+
         ret = {
-            'service_type': service_group,
-            'execution_id': execution_id,
-            'dns_names': [s.dns_name for s in services]
+            'uid': uid,
+            'role': role
         }
 
         self.write(ret)