+++ zoe_api/web/utils.py
@@ -15,13 +15,13 @@
 
 """Functions needed by the Zoe web interface."""
 
-import base64
 import logging
 
 from zoe_lib.config import get_conf
 
 from zoe_api.auth.base import BaseAuthenticator  # pylint: disable=unused-import
 from zoe_api.auth.ldap import LDAPAuthenticator
+from zoe_api.auth.ldapsasl import LDAPSASLAuthenticator
 from zoe_api.auth.file import PlainTextAuthenticator
 import zoe_api.exceptions
 from zoe_api.web.custom_request_handler import ZoeRequestHandler
@@ -54,26 +54,18 @@ def catch_exceptions(func):
 
 
 def missing_auth(handler: ZoeRequestHandler):
-    """Sends a 401 response that enables basic auth"""
-    handler.set_status(401, 'Could not verify your access level for that URL. You have to login with proper credentials.')
-    handler.set_header('WWW-Authenticate', 'Basic realm="Login Required"')
-    handler.finish()
+    """Redirect to login page."""
+    handler.redirect(handler.get_argument('next', u'/login'))
 
 
-def get_auth(handler: ZoeRequestHandler):
-    """Try to authenticate a request."""
-
-    auth_header = handler.request.headers.get('Authorization')
-    if auth_header is None or not auth_header.startswith('Basic '):
-        raise zoe_api.exceptions.ZoeAuthException
-
-    auth_decoded = base64.decodebytes(bytes(auth_header[6:], 'ascii')).decode('utf-8')
-    username, password = auth_decoded.split(':', 2)
-
+def get_auth_login(username, password):
+    """Authenticate username and password against the configured user store."""
     if get_conf().auth_type == 'text':
         authenticator = PlainTextAuthenticator()  # type: BaseAuthenticator
     elif get_conf().auth_type == 'ldap':
-        authenticator = LDAPAuthenticator()
+        authenticator = LDAPAuthenticator()  # type: BaseAuthenticator
+    elif get_conf().auth_type == 'ldapsasl':
+        authenticator = LDAPSASLAuthenticator()  # type: BaseAuthenticator
     else:
         raise zoe_api.exceptions.ZoeException('Configuration error, unknown authentication method: {}'.format(get_conf().auth_type))
     uid, role = authenticator.auth(username, password)
@@ -81,6 +73,18 @@ def get_auth(handler: ZoeRequestHandler):
         raise zoe_api.exceptions.ZoeAuthException
 
     return uid, role
+
+
+def get_auth(handler: ZoeRequestHandler):
+    """Try to authenticate a request."""
+
+    if handler.get_secure_cookie('zoe'):
+        cookie_val = str(handler.get_secure_cookie('zoe'))
+        uid, role = cookie_val[2:-1].split('.')
+        log.info('Authentication done using cookie')
+        return uid, role
+    else:
+        handler.redirect(handler.get_argument('next', u'/login'))
 
 
 def error_page(handler: ZoeRequestHandler, error_message: str, status: int):